Bug Report: https://trello.com/c/XcNX9YeH/40-bug-fixing-geburtstag
Replication: Selecting a date twice in a row.

Identified Problem: Date picker is being called with onFocusListener on the date EditText view. After setting the date, focus didn`t go away. Selecting a already focused editText was causing problems( not sure why).

Implemented Work Around: After Setting a date the focus is passed to the first focusable view from the top. At the moment it kind of fits, but it`s not a perfect solution. Crash avoidence.