https://trello.com/c/sWyu4OgI/70-change-the-the-text-of-the-first-question-in-the-sis-method
Identified Problem:
The Questionnaire was missing one question. After being told the three words the user has to be asked to repeat them immediately.

Implemented Solution:
Changed the first question text. Now after 3 words are told, the user is asked to repeat them before going to the next questions. The last question to repeat them again is still there.

The user has to mention at least one of the three words to be able to go to the next question. If none of the words mentioned, the question with the three words are repeated and the user is asked to name them again and again until he names at least one word from the list.

Note: currently in German, the words have to be repeated with the Artikel.