Movement Recognition fix to not stay for ever after being recognized.
Issue:
https://trello.com/c/hUF3v5Pb/72-correct-the-following-behaviour-once-gehf%C3%A4fig-for-ever-gehf%C3%A4hig

Identified Problem:
Step event time was in different time format than System.nanoTime() which was used to check their validity. So steps were valid for a very long time.

Implemented solution:
Now both the event and the checks on saved event times will use System.nanoTime(). This could recorded step event time a few nanoseconds later, but it should be fine for our purpose.

Increased the steps timestamps buffer size from 9 to 50, which should allow more new steps to be in the array and avoid being full and avoid not recording new steps.

Notes:
Each step is valid for 10sec, and they are checked for validity every 5 seconds. 3 valid, recognized steps are needed for a state of "Walking - Gehfähigkeit".